fix: [#55135] fix initial date in datepicker
The implementation of datepickers respecting a unit's complete validity introduced a "bug"
that has been in MO for a long time. The reasoning being that a date-value never was populated,
which we changed to now being populated with "end-date", this triggered some logic which
then set the "end-date"-picker's value to being the unit's end-date. I'm not completely sure
if this fix will result in other functions to fail, but the datepickers now always
use the internal value