feat!: [#60942] serve frontend on /old/
BREAKING CHANGE:

Prior to this commit, the vue frontend assumed it was running on the
root path, i.e. that links to `/medarbejder/` or `/` would work. As a
step towards migrating to the new naldo frontend, we are switching the
old vue frontend from `/` to `/old/` and the new naldo frontend from
`/new/` to `/`.

Please ensure that your load balancer is updated before updating.